requirements.txt
pytest
coverage
mysql-connector-python
phpserialize